+++ src/frontend/web_application/src/modules/pi/services/pi/index.ts
@@ -1,3 +1,5 @@
+import { PI, MessagePI } from '../../types';
+
 export const PI_PROPERTIES = ['comportment', 'technic', 'context'];
 export const PI_MESSAGE_PROPERTIES = ['content', 'transport', 'social'];
 export const PI_LEVEL_DISABLED = 'disabled-pi';
@@ -6,7 +8,7 @@ export const PI_LEVEL_BAD = 'bad';
 export const PI_LEVEL_GOOD = 'good';
 export const PI_LEVEL_SUPER = 'super';
 
-export const getPiClass = (piAggregate) => {
+export const getPiClass = (piAggregate: number | any) => {
   if (Number.isNaN(piAggregate)) return PI_LEVEL_DISABLED;
 
   if (piAggregate < 25) return PI_LEVEL_UGLY;
@@ -24,7 +26,7 @@ export const getAngles = () => {
   return 360 / piLength;
 };
 
-export const getAveragePI = (pi, piProps = PI_PROPERTIES) => {
+export const getAveragePI = (pi: PI | MessagePI | void, piProps: string[]) => {
   if (!pi) return NaN;
 
   return Math.round(
@@ -34,3 +36,23 @@ export const getAveragePI = (pi, piProps = PI_PROPERTIES) => {
 
 export const getAveragePIMessage = ({ message }) =>
   getAveragePI(message.pi_message, PI_MESSAGE_PROPERTIES);
+
+export type ScoreLetter = 'A' | 'B' | 'C' | 'D' | 'E';
+
+export const computeScoreLetter = (percent: number) => {
+  let key = Math.ceil(percent / 20);
+  if (key < 1) {
+    key = 1;
+  }
+  if (key > 5) {
+    key = 5;
+  }
+
+  // reverse: A eq 0; E eq 4;
+  const baseCharCode = 5 - key;
+
+  return String.fromCharCode(65 + baseCharCode) as ScoreLetter;
+};
+
+export const computePiScoreLetter = (pi: PI) =>
+  computeScoreLetter(getAveragePI(pi, PI_PROPERTIES));